backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg[binary]==3.2.4
django-environ>=0.11
gunicorn>=22.0
